JSpikeStack changes now integrated into JAER, and all should build nicely now.
git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@3533 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/jaer/projects/neuralnets/SpikeStackArrayFilter.java
@@ -150,8 +150,8 @@ public class SpikeStackArrayFilter extends EventFilter2D{
         VisualMapper R=new VisualMapper();
         R.inDimX=(short)chip.getSizeX();
         R.inDimY=(short)chip.getSizeY(); 
-        R.outDimX=net.lay(0).dimx;
-        R.outDimY=net.lay(0).dimy;
+        R.outDimX=(short)net.lay(0).dimx;
+        R.outDimY=(short)net.lay(0).dimy;
         return R;
     }
